Split private Prophet impls across submodules to make it easier to navigate
