+++ crates/augurs-prophet/Cargo.toml
@@ -13,7 +13,6 @@ itertools.workspace = true
 num-traits.workspace = true
 rand = "0.8.5"
 thiserror.workspace = true
-tracing.workspace = true
 
 [dev-dependencies]
 augurs-testing.workspace = true
