+++ crates/augurs-prophet/src/lib.rs
@@ -32,6 +32,6 @@ pub use prophet::{
     options::{
         EstimationMode, GrowthType, OptProphetOptions, ProphetOptions, Scaling, SeasonalityOption,
     },
-    FeaturePrediction, IncludeHistory, Predictions, Prophet,
+    predict::{FeaturePrediction, IncludeHistory, Predictions},
+    Prophet,
 };
-use util::FloatIterExt;
